+++ FractalComplex/FractalComplex/Camera.h
@@ -5,11 +5,11 @@ class Camera{
 public:
 	Vector2d pos;
 	Vector2d border;
-	double angle = 0;
-	double scale = 30;
-	double scaleVel = 2;
+	long double angle = 0;
+	long double scale = 30;
+	long double scaleVel = 2;
 	Camera();
-	Camera(Vector2d _pos, Vector2d _border, double _scale);
+	Camera(Vector2d _pos, Vector2d _border, long double _scale);
 	Vector2d transform(Vector2d p);
 	~Camera();
 };
